requirements.txt
Django==4.0
psycopg2-binary==2.8.5
django-mptt==0.14.0
django-bootstrap4==22.3

